Usuwanie zdjęć po usunięciu oferty, drobne zmiany

+++ php_scripts/user_panel/delete_offer.php
@@ -11,8 +11,21 @@ try{//usuwanie zdjęć trzeba tu zrobić
     require_once '../connect.php';
     $connection = mysqli_connect($host,$db_user,$db_password,$db_name);
     if($connection->connect_errno==0){
-        $sql="DELETE FROM users_offers WHERE seller='$user' AND offer_id='$offer'";
-        $result=mysqli_query($connection,$sql);
+        $sql="SELECT photo1, photo2 FROM users_offers WHERE seller='$user' AND offer_id='$offer'";
+        $result = mysqli_query($connection,$sql); 
+        if($result){
+            $result=mysqli_fetch_row($result);
+            foreach($result as $photo){
+                if($photo){
+                    $path="../../$photo";
+                    if(file_exists($path)){
+                        unlink($path);
+                    }
+                }
+            }
+            $sql="DELETE FROM users_offers WHERE seller='$user' AND offer_id='$offer'";
+            $result=mysqli_query($connection,$sql);
+        }
         mysqli_close($connection);
     }
 }
